background.js
This javascript file handles message passing. background.js is not called within popup.html, instead it communicates with popup.js in the "background", hence why I titled it as such. When the button in your popup is clicked, it sends a message to the background script, background.js, to apply the gradient.